==> ml7/ds_competition/2.data_split.ipynb <==
import pandas as pd 
import warnings 
warnings.filterwarnings("ignore")
card =  pd.read_csv("../input/cleaned_card.csv",index_col='学生id',encoding='utf-8')
score =  pd.read_csv("../input/cleaned_score.csv",index_col='学生id',encoding='utf-8')
data = pd.concat([score,card],axis =1).drop('学院编号', 1)

data.head(10)

data.isnull().sum()

# 使用均值替换nan
data.fillna(data.mean(),inplace=True)
data.isnull().sum()

ID_test =  pd.read_csv("../input/studentID_test.csv",encoding='utf-8')
ID_sub_train = pd.read_csv("../input/subsidy_train.csv",encoding='utf-8')

ID_sub_train.head()

X_test = data.loc[ID_test.学生id,:]
X_train = data.loc[ID_sub_train.学生id,:]
y_train = ID_sub_train.助学金金额

print(X_test.shape,X_train.shape,y_train.shape)

X_test.to_pickle('../input/X_test.pkl')
X_train.to_pickle('../input/X_train.pkl')
y_train.to_pickle('../input/y_train.pkl')
