==> ml7/ds_competition/4.RandomForest_and_GBDT.ipynb <==
import pandas as pd
import numpy as np
from sklearn.ensemble import RandomForestClassifier
from sklearn.ensemble import GradientBoostingClassifier
import xgboost as xgb
#加载数据
TRAINING = pd.read_csv("../input/train.csv")
TESTING = pd.read_csv("../input/test.csv")

TRAINING.head()

feature_list = TESTING.columns.values.tolist()
X_train = TRAINING[feature_list]
y_train = TRAINING['sub']
X_test = TESTING[feature_list]

y_train.head()

import warnings 
warnings.filterwarnings("ignore")
from sklearn.linear_model import LogisticRegression
from sklearn.model_selection import GridSearchCV

# 一个训练优化模型的函数
def train_optimal_classifier(clf,param):
    grid_search = GridSearchCV(
        clf,
        param_grid=param,
        cv=5,
        verbose=0,
        scoring='f1_weighted',
        error_score=0,
        refit=True ,
        n_jobs=-1
        )
    grid_search.fit(X_train, y_train)
    print("Best parameters")
    print(grid_search.best_params_)
    print('best F1: ')
    print(grid_search.best_score_)
    return (grid_search.best_estimator_)


clf1 = RandomForestClassifier()
param1 = {  "n_estimators":[10,50,100,200,400],
             }
train_optimal_classifier(clf1,param1)

clf1 = RandomForestClassifier(n_estimators=20)
clf1.fit(X_train, y_train)
pred = clf1.predict(X_test)

test_result = pd.DataFrame(columns=["subsidy"])
test_result.subsidy = pred
test_result.subsidy = test_result.subsidy.apply(lambda x:int(x))
test_result.to_csv("../data3.csv",index=False)

X_test
